backend/requirements.txt
# Core framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.27
alembic==1.13.1

# Celery / Redis
celery==5.3.6
redis==5.0.2

# Document processing
PyMuPDF==1.23.22
PyPDF2==3.0.1
python-docx==1.1.0
pytesseract==0.3.10
Pillow==10.2.0

# Data handling
numpy<2.0
pandas==2.2.0
pyarrow==15.0.0

# RAG — embeddings and vector search
# FAISS + PyTorch (GPU) are installed separately in Dockerfile
sentence-transformers==3.0.1

# HTTP client (async)
httpx==0.27.0

# Encryption
cryptography==42.0.5

# PubMed (fallback — primary uses httpx)
biopython==1.83

# Utilities
werkzeug==3.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4

# Logging
python-json-logger==2.0.7
